backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-pymongo==2.3.0
python-dotenv==0.19.0
PyJWT==2.8.0
Werkzeug==2.0.1
pymongo==3.12.0
phonenumbers==8.13.30
bcrypt==4.0.1 